requirements.txt
lxml
scrapy
beautifulsoup4
pymongo
pillow
requests
python-dateutil